backend/requirements.txt
asgiref
dj-database-urls
Django
django-cors-headers
djangorestframework
djangorestframework_simplejwt
gunicorn
packaging
psycopg2-binary
PyJWT
sqlparse
tzdata
waitress
whitenoise
